[FIX] pos_product_multi_barcode: incoherent data
If a `barcode` is already in `product_by_barcode`, do not override
it, as Odoo does with `product_by_id`. This can lead to incoherent
data if something changes after product is first loaded and then
reloads (if limited loading is enabled), or it is modified in
`product_by_id` and not in `product_by_barcode`.